fo-db-main/server/requirements.txt
pyinstaller
Flask
SQLAlchemy
waitress
pyOpenSSL
